# openapi/swagger.yaml
openapi: 3.0.1
info:
  title: Process management
  description: Process management
  version: 1.0.0
servers:
- url: /
paths:
...
  /boxStatus/{id}:
    get:
      tags:
        - BoxStatus
      operationId: getBoxStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoxStatus'
    post:
      tags:
        - BoxStatus
      operationId: createBoxStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoxStatus'
    put:
      tags:
        - BoxStatus
      operationId: updateBoxStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoxStatus'
    delete:
      tags:
        - BoxStatus
      operationId: deleteBoxStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
...
  /chargingStationStatus/{id}:
    get:
      tags:
        - ChargingStationStatus
      operationId: getChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    post:
      tags:
        - ChargingStationStatus
      operationId: createChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    put:
      tags:
        - ChargingStationStatus
      operationId: updateChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargingStationStatus'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChargingStationStatus'
    delete:
      tags:
        - ChargingStationStatus
      operationId: deleteChargingStationStatus
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: efeu ID
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    BoxStatus:
      required:
        - id
        - state
        - latitude
        - longitude
        - boxLoad
        - timestamp
      type: object
      properties:
        id:
          type: string
        boxState:
          type: string
          enum:
            - Free
            - InUse
            - Dirty
            - Disrupted
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        boxLoad:
          type: string
          enum:
            - Empty
            - Full
        timestamp:
          type: string
          format: date-time
        vehicleId:
          type: string
        mountId:
          type: string
...
    ChargingStationStatus:
      required:
        - id
        - state
        - timestamp
      type: object
      properties:
        id:
          type: string
        chargingStationState:
          type: string
          enum:
            - Free
            - InUse
            - Disrupted
        vehicleId:
          type: string
        timestamp:
          type: string
          format: date-time